backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
torch>=2.2.0
pillow>=10.3.0
pydantic>=2.7.0
numpy>=1.26.0
python-multipart>=0.0.9
segment-anything>=1.0
pyyaml>=6.0

